Restructure and reorganize api code
