+++ apps/api/src/controllers/index.ts
@@ -1,0 +1,9 @@
+import * as auth from './auth';
+import * as user from './user';
+
+const controllers = {
+  auth,
+  user,
+};
+
+export default controllers;
