+++ apps/api/src/database/index.ts
@@ -1,0 +1,5 @@
+import { Pool } from 'pg';
+
+export const pool = new Pool({
+  connectionString: process.env.DATABASE_URL,
+});
