+++ apps/api/src/routes/index.ts
@@ -1,0 +1,9 @@
+import auth from './auth';
+import user from './user';
+
+const routes = {
+  auth,
+  user,
+};
+
+export default routes;
